# 8313238: Improve the performance of strlen for null-terminated strings