--- a/src/java.base/share/classes/jdk/internal/foreign/StringSupport.java
+++ b/src/java.base/share/classes/jdk/internal/foreign/StringSupport.java
@@ -25,12 +25,12 @@
 
 package jdk.internal.foreign;
 
-import jdk.internal.foreign.abi.SharedUtils;
+import jdk.internal.util.ArraysSupport;
 
+import java.lang.foreign.FunctionDescriptor;
+import java.lang.foreign.Linker;
 import java.lang.foreign.MemorySegment;
-import java.lang.invoke.MethodHandles;
-import java.lang.invoke.VarHandle;
-import java.nio.ByteOrder;
+import java.lang.invoke.MethodHandle;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 
@@ -40,6 +40,20 @@ import static java.lang.foreign.ValueLayout.*;
  * Miscellaneous functions to read and write strings, in various charsets.
  */
 public class StringSupport {
+
+    private static final MethodHandle STRLEN;
+    private static final MethodHandle STRLEN_CRITICAL;
+    private static final long MAX_CRITICAL_SIZE = 1024;
+
+    static {
+        Linker linker = Linker.nativeLinker();
+        var strlen = linker.defaultLookup().find("strlen").orElseThrow();
+        var description = FunctionDescriptor.of(JAVA_LONG, ADDRESS);
+
+        STRLEN = linker.downcallHandle(strlen, description);
+        STRLEN_CRITICAL = linker.downcallHandle(strlen, description, Linker.Option.isTrivial());
+    }
+
     public static String read(MemorySegment segment, long offset, Charset charset) {
         return switch (CharsetKind.of(charset)) {
             case SINGLE_BYTE -> readFast_byte(segment, offset, charset);
@@ -58,10 +72,7 @@ public class StringSupport {
         }
     }
     private static String readFast_byte(MemorySegment segment, long offset, Charset charset) {
-        long len = segment.byteSize() < Long.MAX_VALUE
-                // We can only read in chunks if the segment is bound
-                ? chunked_strlen_byte(segment, offset)
-                : strlen_byte(segment, offset);
+        long len = native_strlen_byte(segment, offset);
         byte[] bytes = new byte[(int)len];
         MemorySegment.copy(segment, JAVA_BYTE, offset, bytes, 0, (int)len);
         return new String(bytes, charset);
@@ -99,78 +110,23 @@ public class StringSupport {
         segment.set(JAVA_INT, offset + bytes.length, 0);
     }
 
-    // Create an array handle for which the index parameter is always zero
-    private static final VarHandle LONG_HANDLE =
-            MethodHandles.insertCoordinates(MethodHandles.byteArrayViewVarHandle(long[].class, ByteOrder.nativeOrder()), 1, 0);
-
-    // This method is inspired by the `glibc/string/strlen.c` implementation
-    private static int chunked_strlen_byte(MemorySegment segment, long start) {
-
-        long startAddress = segment.address() + start;
-        int headCount = (int)(SharedUtils.alignUp(startAddress, Long.BYTES) - startAddress);
-
-        int offset = 0;
-        for (; offset < headCount; offset++) {
-            byte curr = segment.get(JAVA_BYTE, start + offset);
-            if (curr == 0) {
-                return offset;
+    private static int native_strlen_byte(MemorySegment segment, long start) {
+        try {
+            if (start > 0) {
+                segment = segment.asSlice(start);
             }
-        }
+            long len = segment.byteSize() < MAX_CRITICAL_SIZE
+                    ? (long)STRLEN_CRITICAL.invokeExact(segment)
+                    : (long)STRLEN.invokeExact(segment);
 
-        // We are now on a long-aligned boundary
-        int bodyCount = (int)Math.min(
-                // Make sure we do not wrap around
-                Integer.MAX_VALUE - Long.BYTES,
-                // Remaining bytes to consider
-                (segment.byteSize() - start - headCount)  // segment.byteSize() might be Long.MAX_VALUE
-        ) & -Long.BYTES;
-
-        for (; offset < bodyCount; offset += Long.BYTES) {
-            // We know we are `long` aligned so, we can save on alignment checking here
-            long curr = segment.get(JAVA_LONG_UNALIGNED, start + offset);
-            // Is this a candidate?
-            if (mightContainZeroByte(curr)) {
-                byte[] arr = new byte[Long.BYTES];
-                // Check the actual content
-                LONG_HANDLE.set(arr, curr);
-                for (int j = 0; j < 8; j++) {
-                    if (arr[j] == 0) {
-                        return offset + j;
-                    }
-                }
+            // On platforms where `size_t` maps to an `int`, we must check if `len < 0`
+            if (len < 0 || len > ArraysSupport.SOFT_MAX_ARRAY_LENGTH) {
+                throw new IllegalArgumentException("String too large");
             }
+            return (int)len;
+        } catch (Throwable e) {
+            throw new IllegalArgumentException(e);
         }
-
-        // Handle the tail
-        return offset + strlen_byte(segment, start + offset);
-    }
-
-    /* Bits 63 and N * 8 (N = 1..7) of this number are zero.  Call these bits
-       the "holes".  Note that there is a hole just to the left of
-       each byte, with an extra at the end:
-
-       bits:  01111110 11111110 11111110 11111110 11111110 11111110 11111110 11111111
-       bytes: AAAAAAAA BBBBBBBB CCCCCCCC DDDDDDDD EEEEEEEE FFFFFFFF GGGGGGGG HHHHHHHH
-
-       The 1-bits make sure that carries propagate to the next 0-bit.
-       The 0-bits provide holes for carries to fall into.
-    */
-    private static final long HI_MAGIC = 0x8080_8080_8080_8080L;
-    private static final long LO_MAGIC = 0x0101_0101_0101_0101L;
-
-    static boolean mightContainZeroByte(long l) {
-        return ((l - LO_MAGIC) & (~l) & HI_MAGIC) != 0;
-    }
-
-    private static int strlen_byte(MemorySegment segment, long start) {
-        // iterate until overflow (String can only hold a byte[], whose length can be expressed as an int)
-        for (int offset = 0; offset >= 0; offset++) {
-            byte curr = segment.get(JAVA_BYTE, start + offset);
-            if (curr == 0) {
-                return offset;
-            }
-        }
-        throw new IllegalArgumentException("String too large");
     }
 
     private static int strlen_short(MemorySegment segment, long start) {