@minborg 2023-08-03
Here are some benchmarks for various string lengths:

Before:

```
Benchmark                         (size)  Mode  Cnt    Score   Error  Units
InternalStrLen.currentSingleByte       4  avgt   30    2.989 ? 0.013  ns/op
InternalStrLen.currentSingleByte     251  avgt   30  124.791 ? 0.210  ns/op
```

After:

```
Benchmark                         (size)  Mode  Cnt    Score   Error  Units
InternalStrLen.nativeSingleByte        4  avgt   30    3.345 ? 0.020  ns/op
InternalStrLen.nativeSingleByte      251  avgt   30    8.753 ? 0.022  ns/op
```